+++ sample/ReadEventWrite/main.go
@@ -39,7 +39,7 @@ func main() {
 
 	// Set the http file server.
 	http.Handle("/", http.StripPrefix("/", http.FileServer(http.Dir("public"))))
-	http.Handle("/dist/", http.StripPrefix("/dist/", http.FileServer(http.Dir("../client/dist"))))
+	http.Handle("/dist/", http.StripPrefix("/dist/", http.FileServer(http.Dir("../../client/dist"))))
 
 	// Start the http server.
 	err := http.ListenAndServe(ListenAddress, nil)
@@ -54,29 +54,12 @@ func onNewSocket(s *glue.Socket) {
 		log.Printf("socket closed with remote address: %s", s.RemoteAddr())
 	})
 
-	// Run the read loop in a new goroutine.
-	go readLoop(s)
+	// Set a function which is triggered during each received message.
+	s.OnRead(func(data string) {
+		// Echo the received data back to the client.
+		s.Write(data)
+	})
 
 	// Send a welcome string to the client.
 	s.Write("Hello Client")
-}
-
-func readLoop(s *glue.Socket) {
-	for {
-		// Wait for available data.
-		// Optional: pass a timeout duration to read.
-		data, err := s.Read()
-		if err != nil {
-			// Just return and release this goroutine if the socket was closed.
-			if err == glue.ErrSocketClosed {
-				return
-			}
-
-			log.Printf("read error: %v", err)
-			continue
-		}
-
-		// Echo the received data back to the client.
-		s.Write(data)
-	}
 }