+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
+## 1.1.1 - 2015-07-21
+### Added
+ - Added socket OnRead event. Either use Read or OnRead.
+ - Added internal read handler with locking...
+ - Updated README.
+
+
 ## 1.1.0 - 2015-07-15
 ### Added
  - Updated TODO.